Match and link __mem
Former-commit-id: fb1a503ccef3d883f37d6542d08a38acbfde8771

+++ libc/string.h
@@ -10,7 +10,7 @@ extern "C" {
 #pragma section code_type ".init"
 void* memcpy(void* dst, const void* src, size_t n);
 void* memset(void* dst, int val, size_t n);
-void __fill_mem(void* dst, int val, unsigned long n);
+void __fill_mem(void* dst, int val, size_t n);
 #pragma section code_type
 
 size_t strlen(const char* s);
